Add JWT Token Authentication (#38)
* Add JWT Token Authentication

This adds JWT token authentication to Islandora. We send these tokens
along with broadcast messages, and can use them to authenticate with
Fedora as well as call back into Islandora.

* Updated the version for JWT in the composer file.

* JwtEventSubscriber tests.

* Fixing copy/pasta mistake in tests (#2)

* Update JWT Structure to Eliminate Nesting

No java JWT parsing libraries like when a JWT contains nested
structures, even when that is allowed in the standard. This
commit updates our code so we put the drupal data at the
root level of the JWT claims, instead of in a subclaim.

* Coding standards for new sniffs